requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Flask==2.3.0
Flask-CORS==4.0.0
